fix: file and commponents
